Fixes #2040
* Change default background to a gradient to darker shade of blue
* Relaxed constraints of lozenges
* Change admin module to simplify image upload
