+++ magda-gateway/views/_head.ejs
@@ -22,6 +22,6 @@
 <body>
   <div class="outer">
     <div class="inner">
-      <h1>MAGDA ADMIN</h1>
+      <h1>MAGDA</h1>
     </div>
   </div>
