+++ magda-gateway/views/admin.ejs
@@ -11,6 +11,18 @@
 
 <%- include( 'admin.js' ) %>
 
+<%- include( 'admin_common.js' ) %>
+
+<%- include( 'admin_me.js' ) %>
+<%- include( 'admin_logo.js' ) %>
+<%- include( 'admin_navigation.js' ) %>
+<%- include( 'admin_pages.js' ) %>
+<%- include( 'admin_stories.js' ) %>
+<%- include( 'admin_user_accounts.js' ) %>
+<%- include( 'admin_language.js' ) %>
+<%- include( 'admin_connectors.js' ) %>
+<%- include( 'admin_home.js' ) %>
+
 </script>
 
 <%- include( '_tail.ejs' ) %>
